Split out Unix permission tests to their own file
